+++ src/entities/form/ui/PictureOption/index.tsx
