+++ src/views/form/ui/FormEditor/index.tsx
@@ -2,11 +2,11 @@
 
 import { useFieldArray, useForm } from 'react-hook-form';
 import { toast } from 'react-toastify';
-import { CreateFormButton } from '@/entities/form/create';
+import { CreateFormButton } from '@/entities/form';
 import { handleFormErrors } from '@/shared/model/formErrorUtils';
 import { FormValues } from '@/shared/types/form/create/type';
 import { Button, PageHeader } from '@/shared/ui';
-import FormContainer from '@/widgets/form/create/ui/FormContainer';
+import FormContainer from '@/widgets/form/ui/FormContainer';
 import { getFormTitle } from '../../model/getFormTitle';
 import { selectOptionData } from '../../model/selectOptionData';
 import { useSubmitForm } from '../../model/useSubmitForm';
